Rename to `/subscriptions`
