+++ backend/src/routes/v1/mod.rs
@@ -1,6 +1,6 @@
 pub mod auth;
 pub mod media;
-pub mod notifications;
+pub mod subsriptions;
 
 use aide::axum::{
     routing::{delete, get, post},
@@ -20,7 +20,7 @@ pub fn handler() -> ApiRouter {
             post(media::create_presigned_upload_url),
         )
         .api_route("/media/config", get(media::get_media_config))
-        .api_route("/notifications", delete(notifications::unsubscribe))
+        .api_route("/subscriptions", delete(subsriptions::unsubscribe))
         .layer(middleware::from_fn(auth_middleware));
 
     public_routes.merge(protected_routes)
